requirements.txt
langchain
openai
requests
python-dotenv
langchain-openai
langchain-community
streamlit
filetype